Changed to Correct White Screen and Close Button error
1) For close button error, required elevated trust in the OOB settings
Error on screen was:
access to the property or method is not allowed unless the application
has elevated permissions, or the code was called through a user
initiated action

2) For white screen issue: increased ClientTimeout from 1200 to 100000
per some SO responses. This controls the timeout between Slverlighht and
the middle tier.  Also remmed out the Focus event in ProductDetail.cs

3) Increased maxReceivedMessageSize in the service behaviors. Just found
it and thought it might help.

4) Turned off trace sensitive